# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Goggins Habit Tracker API
  description: Backend API for the Goggins Habit Tracker application, providing authentication, data persistence, and AI proxies.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskDifficulty:
      type: string
      enum: [Easy, Medium, Hard, Savage]
    Task:
      type: object
      required: [id, date, description, difficulty, completed, category, estimatedTime]
      properties:
        id:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        difficulty:
          $ref: '#/components/schemas/TaskDifficulty'
        completed:
          type: boolean
        category:
          type: string
        story:
          type: string
        recurringMasterId:
          type: string
        estimatedTime:
          type: number
        actualTime:
          type: number
          nullable: true
        goalAlignment:
          type: number
        alignedGoalId:
          type: string
        justification:
          type: string
        time:
          type: string
        betAmount:
          type: number
        betMultiplier:
          type: number
        betPlaced:
          type: boolean
        betWon:
          type: boolean
        recurrenceRule:
          type: string
          enum: [None, Daily, Weekly, Weekdays, Weekends]
          description: "Added for convenience in API, though frontend separates Task/RecurringTask"
    RecurringTask:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        difficulty:
          $ref: '#/components/schemas/TaskDifficulty'
        category:
          type: string
        recurrenceRule:
          type: string
          enum: [Daily, Weekly, Weekdays, Weekends]
        startDate:
          type: string
        estimatedTime:
          type: number
        goalAlignment:
          type: number
        alignedGoalId:
          type: string
        justification:
          type: string
        time:
          type: string
        # Completions map is complex to model in SQL/flat API, 
        # normally we'd expand these into individual Task instances or a separate completions table.
        # For this API spec, we'll keep the object structure to match frontend for now, 
        # but in a real DB this would likely differ.
        completions: 
          type: object
          additionalProperties:
             type: object
             properties:
                completed: {type: boolean}
                actualTime: {type: number}
                time: {type: string}
                betAmount: {type: number}
                betWon: {type: boolean}
...
    SideQuest:
      type: object
      properties:
        id: {type: string}
        description: {type: string}
        difficulty: { $ref: '#/components/schemas/TaskDifficulty' }
        dailyGoal: {type: number}
        completions:
          type: object
          additionalProperties: {type: number}
...
    Wish:
      type: object
      properties:
        id: {type: string}
        description: {type: string}
        explanation: {type: string}
        label: {type: string}
    CoreTask:
       type: object
       properties:
         id: {type: string}
         description: {type: string}
         explanation: {type: string}
         label: {type: string}

    # --- AI Schemas ---
...
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Get tasks (optional date range filter)
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      tags: [Tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    put:
      tags: [Tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Task'
      responses:
        200:
          description: Updated
          content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Task'
    delete:
      tags: [Tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        204:
          description: Deleted
  /recurring-tasks:
    get:
      tags: [Tasks]
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/RecurringTask'
    post:
      tags: [Tasks]
      requestBody:
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/RecurringTask'
      responses:
        201:
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/RecurringTask'
  /recurring-tasks/{id}:
    put:
      tags: [Tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringTask'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringTask'
    delete:
      tags: [Tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        204: {description: Deleted}
...
  /side-quests:
    get:
      tags: [Tasks]
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SideQuest'
    post:
      tags: [Tasks]
      requestBody:
        content:
          application/json:
             schema:
               $ref: '#/components/schemas/SideQuest'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SideQuest'
  /side-quests/{id}:
    put:
      tags: [Tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
           application/json:
             schema:
               $ref: '#/components/schemas/SideQuest'
      responses:
        200:
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/SideQuest'
    delete:
      tags: [Tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        204: {description: Deleted}
...
  /wish-list:
    get:
      tags: [Tasks]
      responses: {200: {content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Wish'}}}}}}
    post:
      tags: [Tasks]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/Wish'}}}}
      responses: {201: {content: {application/json: {schema: {$ref: '#/components/schemas/Wish'}}}}}
  /wish-list/{id}:
     delete:
       tags: [Tasks]
       parameters: [{in: path, name: id, required: true, schema: {type: string}}]
       responses: {204: {description: Deleted}}
  /core-list:
    get:
      tags: [Tasks]
      responses: {200: {content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/CoreTask'}}}}}}
    post:
      tags: [Tasks]
      requestBody: {content: {application/json: {schema: {$ref: '#/components/schemas/CoreTask'}}}}
      responses: {201: {content: {application/json: {schema: {$ref: '#/components/schemas/CoreTask'}}}}}
  /core-list/{id}:
     delete:
       tags: [Tasks]
       parameters: [{in: path, name: id, required: true, schema: {type: string}}]
       responses: {204: {description: Deleted}}

  # --- AI Services ---
...